requirements.txt
tensorflow
opencv-python
numpy
pillow
streamlit
pytesseract
scikit-learn
